% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_process_subfuncs.R
\name{compute_uc}
\alias{compute_uc}
\title{Compute unmapped concept distributions & per patient medians}
\usage{
compute_uc(
  cohort,
  uc_tbl,
  site_col = "site",
  grouped_list = "site",
  time = FALSE,
  omop_or_pcornet = "omop"
)
}
\arguments{
\item{cohort}{table of cohort members with at least \code{site}, \code{person_id}, \code{start_date}, and \code{end_date}}

\item{uc_tbl}{A table with information about each of the variables that should be examined
in the analysis. This table should contain the following columns:
\itemize{
\item \code{variable} | \emph{character} | a string label for the variable being evaluated for unmapped values
\item \code{domain_tbl} | \emph{character} | the CDM table where the variable is found
\item \code{unmapped_field} | \emph{character} | the name of the field where unmapped values should be identified
\item \code{unmapped_values} | \emph{character} | a string or vector with each value that should be considered as "unmapped"
\item \code{concept_field} | \emph{character} | the string name of the field in the domain table where the concepts are located (if codeset is provided)
\item \code{date_field} | \emph{character} | the name of the field in the domain table with the date that should be used for temporal filtering
\item \code{vocabulary_field} | \emph{character} | for PCORnet applications, the name of the field in the domain table with a vocabulary identifier to differentiate concepts from one another (ex: dx_type); can be set to NA for OMOP applications
\item \code{codeset_name} | \emph{character} | (optional) the name of the codeset that can be used to define a variable of interest (ex: evaluate unit completeness for a specific drug)
\item \code{filter_logic} | \emph{character} | (optional) logic to be applied to the domain_tbl in order to achieve the definition of interest; should be written as if you were applying it in a dplyr::filter command in R
}}

\item{site_col}{the name of the column with site information (either site or site_summ)}

\item{grouped_list}{list of columns that should be used to group the analysis tables}

\item{time}{boolean indicating whether the analysis is being executed over time}

\item{omop_or_pcornet}{string indicating the data model of the underlying CDM data (either omop or pcornet)}
}
\description{
Compute unmapped concept distributions & per patient medians
}
\keyword{internal}
