% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uc_input_file_omop}
\alias{uc_input_file_omop}
\title{UC Sample Input File -- OMOP}
\format{
\subsection{\code{uc_input_file_omop}}{

A data frame with 8 columns
\describe{
\item{variable}{A string label for the variable of interest}
\item{domain_tbl}{The name of the CDM table where the variable can be found}
\item{unmapped_field}{The field in the domain_tbl that should be evaluated for unmapped values}
\item{unmapped_values}{A comma separated string with values that indicate "unmapped" information; NULL values are considered by default.}
\item{concept_field}{The field in the domain_tbl that should be used to join to the codeset; only required when codeset is provided}
\item{date_field}{The date field in the domain_tbl that should be used to filter the dataset to the cohort period and for longitudinal analyses}
\item{codeset_name}{The name of the codeset as found in the specs directory; file extension should not be included}
\item{filter_logic}{optional; a string to be parsed as logic to filter the default_tbl and better identify the variable of interest}
}
}
}
\usage{
uc_input_file_omop
}
\description{
A sample version of the file structure expected for the uc_input_file
parameter in the \code{uc_process} function. The user should recreate this file
structure and include their own variables.
}
\details{
Please note that the codesets should be stored in the \code{file_subdirectory} established
when \code{ssdqa.gen::initialize_dq_session} is executed.

Examples of appropriately structured codeset files are attached to the ssdqa.gen
package and can be accessed with \verb{ssdqa.gen::}
}
\keyword{datasets}
