% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_output_subfuncs.R
\name{uc_ms_exp_la}
\alias{uc_ms_exp_la}
\title{\emph{Multi Site, Exploratory, Longitudinal}}
\usage{
uc_ms_exp_la(
  process_output,
  output_col,
  filter_variable = NULL,
  facet = NULL,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{tabular output from \code{uc_process}}

\item{output_col}{name of the column from process_output to be used in the analysis.
can be any of the proportion or either of the median_site_* columns}

\item{filter_variable}{the name(s) of variables that should be included on the plot}

\item{facet}{fields by which the graph should be facetted}

\item{large_n}{a boolean indicating whether the large N visualization, intended for a high
volume of sites, should be used; defaults to FALSE}

\item{large_n_sites}{a vector of site names that can optionally generate a filtered visualization}
}
\value{
a line graph showing the proportion or median per patient of
unmapped concepts per site & variable across the time series
}
\description{
\emph{Multi Site, Exploratory, Longitudinal}
}
\keyword{internal}
