% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_output_subfuncs.R
\name{uc_ss_anom_la}
\alias{uc_ss_anom_la}
\title{\emph{Single Site, Anomaly Detection, Longitudinal}}
\usage{
uc_ss_anom_la(
  process_output,
  output_col = NULL,
  filter_variable = NULL,
  facet = NULL
)
}
\arguments{
\item{process_output}{tabular output from \code{uc_process}}

\item{output_col}{name of the column from process_output to be used in the analysis.
can be either of the proportion columns}

\item{filter_variable}{the name(s) of variables that should be included on the plot}

\item{facet}{fields by which the graph should be facetted}
}
\value{
if analysis was executed by year or greater, a P Prime control chart
is returned with outliers marked with orange dots

\if{html}{\out{<div class="sourceCode">}}\preformatted{     if analysis was executed by month or smaller, an STL regression is
     conducted and outliers are marked with red dots. the graphs representing
     the data removed in the regression are also returned
}\if{html}{\out{</div>}}
}
\description{
\emph{Single Site, Anomaly Detection, Longitudinal}
}
\keyword{internal}
