% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_output_subfuncs.R
\name{uc_ss_exp_cs}
\alias{uc_ss_exp_cs}
\title{\emph{Single Site, Exploratory, Cross-Sectional}}
\usage{
uc_ss_exp_cs(process_output, output_col, facet = NULL)
}
\arguments{
\item{process_output}{tabular output from \code{uc_process}}

\item{output_col}{name of the column from process_output to be used in the analysis.
can be any of the numeric cols, including either proportion col or any of the
median cols}

\item{facet}{fields by which the graph should be facetted}
}
\value{
a bar graph displaying either the proportion of unmapped rows/patients or the
median number of unmapped values per patient for each variable
}
\description{
\emph{Single Site, Exploratory, Cross-Sectional}
}
\keyword{internal}
