% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_output_subfuncs.R
\name{uc_ss_exp_la}
\alias{uc_ss_exp_la}
\title{\emph{Single Site, Exploratory, Longitudinal}}
\usage{
uc_ss_exp_la(process_output, output_col, facet = NULL)
}
\arguments{
\item{process_output}{tabular output from \code{uc_process}}

\item{output_col}{name of the column from process_output to be used in the analysis.
can be any of the proportion or median columns}

\item{facet}{fields by which the graph should be facetted}
}
\value{
a line graph showing the proportion or median per patient of
unmapped concepts per variable across the time series
}
\description{
\emph{Single Site, Exploratory, Longitudinal}
}
\keyword{internal}
